moved the interaction util files
